+++ src/main/java/net/grid/vampiresdelight/common/CommonSetup.java
@@ -2,7 +2,6 @@ package net.grid.vampiresdelight.common;
 
 import net.grid.vampiresdelight.common.registry.VDItems;
 import net.grid.vampiresdelight.common.registry.VDPotions;
-import net.grid.vampiresdelight.common.registry.VDStats;
 import net.minecraft.world.level.block.DispenserBlock;
 import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
 
@@ -10,7 +9,6 @@ public class CommonSetup {
     public static void init(final FMLCommonSetupEvent event) {
         event.enqueueWork(() -> {
             VDPotions.registerPotionMixes();
-            VDStats.registerModStats();
             registerDispenserBehaviors();
         });
     }