+++ src/main/java/net/grid/vampiresdelight/data/VDDataGenerators.java
@@ -1,11 +1,15 @@
 package net.grid.vampiresdelight.data;
 
-import de.teamlapen.vampirism.core.ModRegistries;
 import net.grid.vampiresdelight.VampiresDelight;
+import net.grid.vampiresdelight.common.world.VDBiomeModifiers;
+import net.grid.vampiresdelight.common.world.VDConfiguredFeatures;
+import net.grid.vampiresdelight.common.world.VDPlacedFeatures;
 import net.grid.vampiresdelight.data.loot.VDBlockLootTables;
 import net.grid.vampiresdelight.data.loot.VDChestLootTables;
 import net.grid.vampiresdelight.data.tag.*;
 import net.minecraft.core.HolderLookup;
+import net.minecraft.core.RegistrySetBuilder;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.data.DataGenerator;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.loot.LootTableProvider;
@@ -15,6 +19,7 @@ import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.common.data.DatapackBuiltinEntriesProvider;
 import net.neoforged.neoforge.common.data.ExistingFileHelper;
 import net.neoforged.neoforge.data.event.GatherDataEvent;
+import net.neoforged.neoforge.registries.NeoForgeRegistries;
 
 import java.util.Collections;
 import java.util.List;
@@ -28,19 +33,24 @@ public class VDDataGenerators {
     public static void gatherData(GatherDataEvent event) {
         DataGenerator generator = event.getGenerator();
         PackOutput output = generator.getPackOutput();
+        CompletableFuture<HolderLookup.Provider> lookupProvider = event.getLookupProvider();
         ExistingFileHelper helper = event.getExistingFileHelper();
+        RegistrySetBuilder registrySetBuilder = new RegistrySetBuilder()
+                .add(Registries.CONFIGURED_FEATURE, VDConfiguredFeatures::createConfiguredFeatures)
+                .add(Registries.PLACED_FEATURE, VDPlacedFeatures::createPlacedFeatures)
+                .add(NeoForgeRegistries.Keys.BIOME_MODIFIERS, VDBiomeModifiers::createBiomeModifiers)
+                .add(Registries.ENCHANTMENT, VDEnchantments::createEnchantments);
 
-        DatapackBuiltinEntriesProvider provider = new DatapackBuiltinEntriesProvider(output, event.getLookupProvider(), ModRegistries.DATA_BUILDER, Set.of(VampiresDelight.MODID));
-        CompletableFuture<HolderLookup.Provider> lookupProvider = provider.getRegistryProvider();
-
-        generator.addProvider(event.includeServer(), new VDRegistries(output, lookupProvider));
         VDBlockTags blockTags = new VDBlockTags(output, lookupProvider, helper);
         generator.addProvider(event.includeServer(), blockTags);
         generator.addProvider(event.includeServer(), new VDItemTags(output, lookupProvider, blockTags.contentsGetter(), helper));
         generator.addProvider(event.includeServer(), new VDBiomeTags(output, lookupProvider, helper));
         generator.addProvider(event.includeServer(), new VDEntityTags(output, lookupProvider, helper));
-        // TODO: Do something with these, it doesn't load
-        //generator.addProvider(event.includeServer(), new VDEnchantmentTags(output, lookupProvider, helper));
+
+        DatapackBuiltinEntriesProvider datapackProvider = new DatapackBuiltinEntriesProvider(output, lookupProvider, registrySetBuilder, Set.of(VampiresDelight.MODID));
+        CompletableFuture<HolderLookup.Provider> builtinLookupProvider = datapackProvider.getRegistryProvider();
+        generator.addProvider(event.includeServer(), datapackProvider);
+        generator.addProvider(event.includeServer(), new VDEnchantmentTags(output, builtinLookupProvider, helper));
 
         generator.addProvider(event.includeServer(), new VDRecipes(output, lookupProvider));
         generator.addProvider(event.includeServer(), new VDDataMaps(output, lookupProvider));
@@ -48,7 +58,7 @@ public class VDDataGenerators {
         generator.addProvider(event.includeServer(), new LootTableProvider(output, Collections.emptySet(), List.of(
                 new LootTableProvider.SubProviderEntry(VDBlockLootTables::new, LootContextParamSets.BLOCK),
                 new LootTableProvider.SubProviderEntry(VDChestLootTables::new, LootContextParamSets.CHEST)
-        ), lookupProvider));
+        ), builtinLookupProvider));
 
         generator.addProvider(event.includeClient(), new VDBlockModels(output, helper));
         VDBlockStates blockStates = new VDBlockStates(output, helper);