+++ src/main/java/net/grid/vampiresdelight/data/loot/VDChestLootTables.java
@@ -1,9 +1,9 @@
 package net.grid.vampiresdelight.data.loot;
 
 import de.teamlapen.vampirism.core.ModBlocks;
-import net.grid.vampiresdelight.data.VDEnchantments;
 import net.grid.vampiresdelight.common.registry.VDItems;
 import net.grid.vampiresdelight.common.registry.VDLootTables;
+import net.grid.vampiresdelight.data.VDEnchantments;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.loot.LootTableSubProvider;
@@ -54,13 +54,11 @@ public class VDChestLootTables implements LootTableSubProvider {
     }
 
     public void lootModifiers(BiConsumer<ResourceKey<LootTable>, LootTable.Builder> consumer) {
-        /*
         consumer.accept(VDLootTables.VD_CHEST_VAMPIRE_DUNGEON, LootTable.lootTable()
                 .withPool(vampiresBiteBookLoot(1, 2))
         );
-         */
         consumer.accept(VDLootTables.VD_CHEST_VAMPIRE_HUT, LootTable.lootTable()
-                //.withPool(vampiresBiteBookLoot(1, 3))
+                .withPool(vampiresBiteBookLoot(1, 3))
                 .withPool(LootPool.lootPool().setRolls(UniformGenerator.between(5, 8))
                         .add(LootItem.lootTableItem(VDItems.ORCHID_COOKIE.get()).setWeight(10))
                         .add(LootItem.lootTableItem(VDItems.ORCHID_TEA.get()).setWeight(3))
@@ -74,14 +72,12 @@ public class VDChestLootTables implements LootTableSubProvider {
                         .add(EmptyLootItem.emptyItem().setWeight(8))
                 )
         );
-        /*
         consumer.accept(VDLootTables.VD_CHEST_ALTAR, LootTable.lootTable()
                 .withPool(vampiresBiteBookLoot(1, 2))
         );
         consumer.accept(VDLootTables.VD_CHEST_CRYPT, LootTable.lootTable()
                 .withPool(vampiresBiteBookLoot(1, 4))
         );
-         */
         consumer.accept(VDLootTables.VD_CHEST_HUNTER_OUTPOST_TENT, LootTable.lootTable()
                 .withPool(LootPool.lootPool().setRolls(UniformGenerator.between(5, 8))
                         .add(LootItem.lootTableItem(VDItems.HARDTACK.get()).setWeight(5))
@@ -102,8 +98,6 @@ public class VDChestLootTables implements LootTableSubProvider {
         );
     }
 
-    // TODO: Currently it can't access enchants from here, not sure why
-    /*
     public LootPool.Builder vampiresBiteBookLoot(int bookWeight, int emptyWeight) {
         HolderLookup.RegistryLookup<Enchantment> lookup = this.provider.lookupOrThrow(Registries.ENCHANTMENT);
         return LootPool.lootPool().setRolls(ConstantValue.exactly(1))
@@ -111,5 +105,4 @@ public class VDChestLootTables implements LootTableSubProvider {
                         .apply((new EnchantRandomlyFunction.Builder()).withEnchantment(lookup.getOrThrow(VDEnchantments.VAMPIRE_BITE))))
                 .add(EmptyLootItem.emptyItem().setWeight(emptyWeight));
     }
-     */
 }